moved all `ChannelConfig` stuff into `node` namespace
