+++ examples/worklet.rs
@@ -1,11 +1,10 @@
 use rand::Rng;
 
 use web_audio_api::alloc::AudioRenderQuantum;
-use web_audio_api::buffer::{ChannelConfig, ChannelConfigOptions};
 use web_audio_api::context::{
     AsBaseAudioContext, AudioContext, AudioContextRegistration, AudioParamId,
 };
-use web_audio_api::node::AudioNode;
+use web_audio_api::node::{AudioNode, ChannelConfig, ChannelConfigOptions};
 use web_audio_api::param::{AudioParam, AudioParamOptions, AutomationRate};
 use web_audio_api::process::{AudioParamValues, AudioProcessor};
 use web_audio_api::SampleRate;
