+++ src/buffer.rs
@@ -1,6 +1,4 @@
 //! General purpose audio signal data structures
-
-use std::sync::atomic::{AtomicU32, AtomicUsize, Ordering};
 use std::sync::Arc;
 
 use crate::alloc::AudioRenderQuantum;
@@ -239,116 +237,6 @@ impl ChannelData {
 
     pub fn as_mut_slice(&mut self) -> &mut [f32] {
         &mut Arc::make_mut(&mut self.data)[..]
-    }
-}
-
-/// How channels must be matched between the node's inputs and outputs.
-#[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub enum ChannelCountMode {
-    /// `computedNumberOfChannels` is the maximum of the number of channels of all connections to an
-    /// input. In this mode channelCount is ignored.
-    Max,
-    /// `computedNumberOfChannels` is determined as for "max" and then clamped to a maximum value of
-    /// the given channelCount.
-    ClampedMax,
-    /// `computedNumberOfChannels` is the exact value as specified by the channelCount.
-    Explicit,
-}
-
-impl From<u32> for ChannelCountMode {
-    fn from(i: u32) -> Self {
-        use ChannelCountMode::*;
-
-        match i {
-            0 => Max,
-            1 => ClampedMax,
-            2 => Explicit,
-            _ => unreachable!(),
-        }
-    }
-}
-
-/// The meaning of the channels, defining how audio up-mixing and down-mixing will happen.
-#[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub enum ChannelInterpretation {
-    Speakers,
-    Discrete,
-}
-
-impl From<u32> for ChannelInterpretation {
-    fn from(i: u32) -> Self {
-        use ChannelInterpretation::*;
-
-        match i {
-            0 => Speakers,
-            1 => Discrete,
-            _ => unreachable!(),
-        }
-    }
-}
-
-/// Options for constructing ChannelConfig
-#[derive(Clone, Debug)]
-pub struct ChannelConfigOptions {
-    pub count: usize,
-    pub mode: ChannelCountMode,
-    pub interpretation: ChannelInterpretation,
-}
-
-impl Default for ChannelConfigOptions {
-    fn default() -> Self {
-        Self {
-            count: 2,
-            mode: ChannelCountMode::Max,
-            interpretation: ChannelInterpretation::Speakers,
-        }
-    }
-}
-
-/// Config for up/down-mixing of channels for audio nodes
-#[derive(Clone, Debug)]
-pub struct ChannelConfig {
-    count: Arc<AtomicUsize>,
-    mode: Arc<AtomicU32>,
-    interpretation: Arc<AtomicU32>,
-}
-
-impl ChannelConfig {
-    /// Represents an enumerated value describing the way channels must be matched between the
-    /// node's inputs and outputs.
-    pub fn count_mode(&self) -> ChannelCountMode {
-        self.mode.load(Ordering::SeqCst).into()
-    }
-    pub fn set_count_mode(&self, v: ChannelCountMode) {
-        self.mode.store(v as u32, Ordering::SeqCst)
-    }
-
-    /// Represents an enumerated value describing the meaning of the channels. This interpretation
-    /// will define how audio up-mixing and down-mixing will happen.
-    pub fn interpretation(&self) -> ChannelInterpretation {
-        self.interpretation.load(Ordering::SeqCst).into()
-    }
-    pub fn set_interpretation(&self, v: ChannelInterpretation) {
-        self.interpretation.store(v as u32, Ordering::SeqCst)
-    }
-
-    /// Represents an integer used to determine how many channels are used when up-mixing and
-    /// down-mixing connections to any inputs to the node.
-    pub fn count(&self) -> usize {
-        self.count.load(Ordering::SeqCst)
-    }
-    pub fn set_count(&self, v: usize) {
-        self.count.store(v, Ordering::SeqCst)
-    }
-}
-
-impl From<ChannelConfigOptions> for ChannelConfig {
-    fn from(opts: ChannelConfigOptions) -> Self {
-        ChannelConfig {
-            count: Arc::new(AtomicUsize::from(opts.count)),
-            mode: Arc::new(AtomicU32::from(opts.mode as u32)),
-            interpretation: Arc::new(AtomicU32::from(opts.interpretation as u32)),
-        }
     }
 }
