+++ src/context.rs
@@ -19,14 +19,15 @@ const LISTENER_NODE_ID: u64 = 1;
 /// listener audio parameters ids are always at index 2 through 12
 const LISTENER_PARAM_IDS: Range<u64> = 2..12;
 
-use crate::buffer::{AudioBuffer, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation};
+use crate::buffer::{AudioBuffer};
 use crate::graph::{NodeIndex, RenderThread};
 use crate::media::{MediaElement, MediaStream};
 use crate::message::ControlMessage;
 use crate::node::{
-    self, AnalyserOptions, AudioBufferSourceNodeOptions, AudioNode, ChannelMergerOptions,
-    ChannelSplitterOptions, ConstantSourceOptions, DelayOptions, GainOptions, IirFilterOptions,
-    PannerOptions, PeriodicWave, PeriodicWaveOptions,
+    self, AnalyserOptions, AudioBufferSourceNodeOptions, AudioNode, ChannelConfigOptions,
+    ChannelCountMode, ChannelInterpretation, ChannelMergerOptions, ChannelSplitterOptions,
+    ConstantSourceOptions, DelayOptions, GainOptions, IirFilterOptions, PannerOptions,
+    PeriodicWave, PeriodicWaveOptions,
 };
 use crate::param::{AudioParam, AudioParamEvent, AudioParamOptions};
 use crate::process::AudioProcessor;
