+++ src/node/audio_buffer.rs
@@ -1,11 +1,12 @@
-use crate::buffer::{AudioBuffer, ChannelConfig, ChannelConfigOptions, Resampler};
+use crate::buffer::{AudioBuffer, Resampler};
 use crate::context::{AsBaseAudioContext, AudioContextRegistration};
 use crate::control::{Controller, Scheduler};
 use crate::media::MediaElement;
 use crate::{SampleRate, RENDER_QUANTUM_SIZE};
 
 use super::{
-    AudioControllableSourceNode, AudioNode, AudioScheduledSourceNode, MediaStreamRenderer,
+    AudioControllableSourceNode, AudioNode, AudioScheduledSourceNode,
+    ChannelConfig, ChannelConfigOptions, MediaStreamRenderer,
 };
 
 /// Options for constructing a AudioBufferSourceNode
