+++ src/node/channel_merger.rs
@@ -1,12 +1,11 @@
 use std::fmt::Debug;
 
 use crate::alloc::AudioRenderQuantum;
-use crate::buffer::{ChannelConfig, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation};
 use crate::context::{AsBaseAudioContext, AudioContextRegistration};
 use crate::process::{AudioParamValues, AudioProcessor};
 use crate::SampleRate;
 
-use super::AudioNode;
+use super::{AudioNode, ChannelConfig, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation};
 
 /// Options for constructing a ChannelMergerNode
 pub struct ChannelMergerOptions {
