+++ src/node/destination.rs
@@ -1,10 +1,9 @@
 use crate::alloc::AudioRenderQuantum;
-use crate::buffer::{ChannelConfig, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation};
 use crate::context::{AsBaseAudioContext, AudioContextRegistration};
 use crate::process::{AudioParamValues, AudioProcessor};
 use crate::SampleRate;
 
-use super::AudioNode;
+use super::{AudioNode, ChannelConfig, ChannelConfigOptions, ChannelCountMode, ChannelInterpretation};
 
 /// Representing the final audio destination and is what the user will ultimately hear.
 pub struct DestinationNode {
